+++ tests/docsrc/tutorial/doc_tutorial_callable.cpp
@@ -192,10 +192,10 @@ the arguments array. The second argument is the number of Variant in the array.
 desc*/
 
 //code
-int myVariadicFunc(const metapp::Variant * arguments, const size_t argumentCount)
+int myVariadicFunc(const metapp::Variant * arguments, const int argumentCount)
 {
 	int total = 0;
-	for(size_t i = 0; i < argumentCount; ++i) {
+	for(int i = 0; i < argumentCount; ++i) {
 		total += arguments[i].cast<int>().get<int>();
 	}
 	return total;
